backend/requirements.txt
Django==4.2.1
djangorestframework==3.16.0
django-filter==25.1
drf-spectacular==0.28.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.10
openai>=1.0.0
pix2text>=0.2.0
pdfkit>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
